Add validation to delete a job code
- You cannot delete a jobcode with RSDs

- You cannot delete a job code with children

+++ api/src/main/kotlin/edu/wgu/osmt/jobcode/JobCodeRepository.kt
@@ -2,6 +2,7 @@ package edu.wgu.osmt.jobcode
 
 import edu.wgu.osmt.api.model.ApiBatchResult
 import edu.wgu.osmt.api.model.JobCodeUpdate
+import edu.wgu.osmt.richskill.RichSkillJobCodeRepository
 import org.jetbrains.exposed.sql.SizedIterable
 import org.jetbrains.exposed.sql.Table
 import org.jetbrains.exposed.sql.and
@@ -21,6 +22,7 @@ interface JobCodeRepository {
     fun findById(id: Long): JobCodeDao?
     fun findByCode(code: String): JobCodeDao?
     fun findByCodeOrCreate(code: String, framework: String? = null): JobCodeDao
+    fun hasChildren(jobCodeDao: JobCodeDao): Boolean
     fun findBlsCode(code: String): JobCodeDao?
     fun create(code: String, framework: String? = null): JobCodeDao
     fun createFromApi(jobCodes: List<JobCodeUpdate>): List<JobCodeDao>
@@ -40,6 +42,10 @@ class JobCodeRepositoryImpl: JobCodeRepository {
     @Autowired
     @Lazy
     lateinit var jobCodeEsRepo: JobCodeEsRepo
+
+    @Autowired
+    @Lazy
+    lateinit var richSkillJobCodeRepository: RichSkillJobCodeRepository
 
     val dao = JobCodeDao.Companion
     override val table = JobCodeTable
@@ -76,6 +82,16 @@ class JobCodeRepositoryImpl: JobCodeRepository {
         return existing ?: create(code, framework)
     }
 
+    override fun hasChildren(jobCodeDao: JobCodeDao): Boolean {
+        return findAll().any { jobCode ->
+            jobCode.code != jobCodeDao.code &&
+            (jobCodeDao.code == JobCodeBreakout.majorCode(jobCode.code) ||
+                    jobCodeDao.code == JobCodeBreakout.minorCode(jobCode.code) ||
+                    jobCodeDao.code == JobCodeBreakout.broadCode(jobCode.code) ||
+                    jobCodeDao.code == JobCodeBreakout.detailedCode(jobCode.code))
+        }
+    }
+
     override fun create(code: String, framework: String?): JobCodeDao {
         val maybeDetailed = JobCodeBreakout.detailedCode(code).let{ findBlsCode(code) }
         return dao.new {
@@ -100,7 +116,7 @@ class JobCodeRepositoryImpl: JobCodeRepository {
     override fun remove(jobCodeId: Long): ApiBatchResult {
         val jobCodeFound = findById(jobCodeId)
         val jobCodeEsFound = jobCodeEsRepo.findById(jobCodeId.toInt())
-        if (jobCodeFound != null && jobCodeEsFound.isPresent) {
+        if (jobCodeFound != null && jobCodeEsFound.isPresent && !hasChildren(jobCodeFound) && richSkillJobCodeRepository.hasRSDs(jobCodeFound)) {
             transaction {
                 table.deleteWhere{ table.id eq jobCodeFound.id }
                 jobCodeEsRepo.delete(jobCodeEsFound.get())
